+++ uvicorn/supervisors/multiprocess.py
@@ -64,7 +64,7 @@ class Process:
 
     def is_alive(self, timeout: float = 5) -> bool:
         if not self.process.is_alive():
-            return False
+            return False  # pragma: full coverage
 
         return self.ping(timeout)
 
@@ -171,7 +171,7 @@ class Multiprocess:
             process.join()
 
             if self.should_exit.is_set():
-                return
+                return  # pragma: full coverage
 
             logger.info(f"Child process [{process.pid}] died")
             process = Process(self.config, self.target, self.sockets)